Changed parsing to work using Long rather than Integer. Changed to actually update the input file.

+++ src/main/java/com/purplefrog/jluadata/LuaParser.java
@@ -381,7 +381,7 @@ public class LuaParser
             }
         }
 
-        return Integer.parseInt(buffer.toString());
+        return Long.parseLong(buffer.toString());
     }
 
     public Double parseRemainingDouble(StringBuilder buffer)
